#446 Update token generation of microsoftapi
